format large balance in home screen

+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/home/HomeViewModel.kt
@@ -171,7 +171,7 @@ class HomeViewModel(
         }
         updateState {
             it.copy(
-                balanceAmount = balanceAmount.toInt().toString(),
+                balanceAmount = balanceAmount.toString(),
                 isBalanceLoading = false
             )
         }
